thematic-analysis-ai-tool/server/requirements.txt
fastapi==0.115.12
passlib==1.7.4
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.1.0
python_jose==3.5.0
Requests==2.32.3
SQLAlchemy==2.0.41
uvicorn==0.34.2
psycopg2-binary==2.9.10
pydantic[email]==2.10.6
cloudinary==1.44.0
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.2
python-multipart==0.0.12
alembic==1.14.0
langchain[openai,anthropic,google-genai,groq]